Create an SFTP connection to Bloomberg Data License

An sftp connection to Bloomberg Datalicense is established. On some Linux systems, this may not work out of the box, as libcurl does not natively support sftp. In that case, you need to compile curl with SFTP support. See here for details: http://askubuntu.com/questions/195545/how-to-enable-sftp-support-in-curl

RblConnect(
  user,
  pw,
  host = "sftp.bloomberg.com",
  port = "22",
  protocol = "sftp",
  verbose = TRUE
)

Arguments

user

The account number assigned by Bloomberg

pw

The password assigned by Bloomberg

host

The connection host

port

The connection port

protocol

The connection protocol

verbose

logical. Should print extra information on failure?

Value

logical. Is the connection succesful?

Examples

if (FALSE) { # \dontrun{
# These are dummy credentials. Replace with the credentials received from Bloomberg 
RblConnect(user = 'dl000000', pw = '0000000000000000') 
} # }